moved EnvironmentModal to shared
