+++ ui/components/designs/lifecycle/DeployStepper.tsx
@@ -33,8 +33,8 @@ import { OPERATION_CENTER_EVENTS } from 'machines/operationsCenter';
 import { capitalize } from 'lodash';
 import FinishFlagIcon from '@/assets/icons/FinishFlagIcon';
 import { DeploymentSummaryFormatter } from './DeploymentSummary';
-import { SEVERITY } from '../../NotificationCenter/constants';
-import EnvironmentModal from '../../General/Modals/EnvironmentModal';
+import { SEVERITY } from '../NotificationCenter/constants';
+import EnvironmentModal from '../shared/Modal/EnvironmentModal';
 import { openViewScopedToDesignInOperator } from '@/utils/utils';
 import { useRouter } from 'next/router';
 import ProviderStoreWrapper from '@/store/ProviderStoreWrapper';
